feat: refactor move api.test utils to test.utils
